Move assertions to separate file and make installation optional
The assertions in question verify invariants of the various data
structures. In this new approach, the assertions are defined separate
from the rest of the source code and retroactively attached to the
respective operation via either TRACE or additional methods on an
existing generic function. The assertions defined in this way are
compiled and attached only if the feature :INCREMENTALIST-DEBUG is
defined at the time when the incrementalist ASDF system is loaded.

Extract and unify assertions into the new approach from two sources:
from the operation definitions and from the file check-wad-graph.lisp.