Fix handling of inherited cells for empty prefix
Before this change, when a wad was moved from the prefix to the
suffix, its list of inherited cells was unconditionally cleared. As a
result, when the prefix consisted of only one wad and that wad was
moved from the prefix to the suffix and then back from the suffix to
the prefix, the wad ended up without a list of inherited cells while
its list of inherited cells should have been the initial reader state
instead. Another failure mode was that removing the last remaining
wad from the prefix (without moving to the suffix) could lead to the
first suffix wad having invalid inherited cells.

After this change, when a wad is moved from the prefix to the suffix,
its list of inherited cells is cleared only if there is still at least
one wad on the prefix. Similarly, if the last remaining wad is removed
from the prefix, the inherited cells of the first wad of the suffix
are set to the initial state of the analyzer.