If necessary, create dummy wads to attach errors to
Due to error recovery, Eclector can skip over token-like input such as
an isolated "." by signaling an error but not producing any parse
results. Before this change, that error had no wad to attach to due to
the lack of parse results. This change adds the creation of dummy
parse results for errors to attach to.