Add fixup processing for circular objects
Before this change, incrementalist was able to produce instances of
LABELED-OBJECT-DEFINITION-WAD and LABELED-OBJECT-REFERENCE-WAD for
sharing like (#1=1 #1#) but not for circularity like #1=(1 #1#). This
change adds methods which customize Eclector's fixup processing so
that cases with circularity produce suitable instances of
LABELED-OBJECT-DEFINITION-WAD and LABELED-OBJECT-REFERENCE-WAD.

Since the added code uses ECLECTOR.READER:NEW-VALUE-FOR-FIXUP,
Incrementalist now depends on Eclector version 0.11 or newer.