test: New test for incomplete labeled objects
The actual problem was with Eclector but the situation has the
potentially to cause problems in Incrementalist as well. Briefly, the
problem arose when a wad representing a labeled object definition had
only children that represented skipped input. In such cases, the
labeled object is never completed and the client has to use Eclector's
RECOVER restart instead.